rename all js files to ts
