+++ src/env.ts
@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   ELASTIC_URL: 'http://localhost:9200/',
   PORT: 3000,
   SECRET_KEY: '__USE_A_SECURE_KEY__',
