+++ src/helpers/auth.ts
@@ -1,11 +1,11 @@
 import * as JWT from 'jsonwebtoken'
-import * as env from '../env'
+import env from '../env'
 
 const signOptions = {
   noTimestamp: true,
 }
 
-export const createToken = (payload) => {
+export const createToken = (payload: Object) => {
   return new Promise((resolve, reject) => {
     JWT.sign(payload, env.SECRET_KEY, signOptions, (err, token) => {
       if (err) {
@@ -16,7 +16,7 @@ export const createToken = (payload) => {
   })
 }
 
-export const verifyToken = (token) => {
+export const verifyToken = (token: string) => {
   return new Promise((resolve, reject) => {
     JWT.verify(token, env.SECRET_KEY, {}, (err, payload) => {
       if (err) {
