+++ src/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "compilerOptions": {
+    "module": "commonjs",
+    "noImplicitAny": true,
+    "noImplicitUseStrict": true,
+    "outDir": "../dist",
+    "strictNullChecks": true,
+    "target": "es2015"
+  }
+}
